web_admin/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn>=21
python-dotenv>=1.0
supabase>=2.4,<3
